Add files via upload
Here is the background information on your task
At GE Aviation, we use the concept of a “Data Lake”.  A Data Lake is a single database instance that contains data from all around Aviation. Everything from financial, delivery (of parts and engines), supplier, engine data, customer data, and so on. 

The advantage of a Data Lake is that allows the developer to make a single connection string to the Data Lake, and as long as the developer has permissions to see the data – they are able to immediately start creating data-driven insights in a centralized repository of data. 

To simplify this experience, we have put all tables into a single schema (database inside the data lake). 

There are 8 datasets that are needed in order to accomplish this task. 

You have been asked by GE Aviation leadership to create a single data set that combines all the data listed above into a single table.   

The data you have been given has sheets related to Flight data (so data relating to the health of the mechanical engine of a plane), Location Data for airports, Manufacturing Data (which relate to various airplane parts), and a manufacturing bill of material (which tells you what engines have been used).

In this final form, we can then use this table to either create a prediction of RUL or visualize the data for easy user consumption. This will be done in task two.